增加一些表设计

+++ internal/db/models/admin_model.go
@@ -7,8 +7,8 @@ type Admin struct {
 	Password  string `field:"password"`  // 密码
 	Fullname  string `field:"fullname"`  // 全名
 	IsSuper   uint8  `field:"isSuper"`   // 是否为超级管理员
-	CreatedAt uint32 `field:"createdAt"` // 创建时间
-	UpdatedAt uint32 `field:"updatedAt"` // 修改时间
+	CreatedAt uint64 `field:"createdAt"` // 创建时间
+	UpdatedAt uint64 `field:"updatedAt"` // 修改时间
 	State     uint8  `field:"state"`     // 状态
 }
