+++ internal/db/models/provider_model.go
@@ -6,8 +6,8 @@ type Provider struct {
 	Username  string `field:"username"`  // 用户名
 	Password  string `field:"password"`  // 密码
 	Fullname  string `field:"fullname"`  // 真实姓名
-	CreatedAt uint32 `field:"createdAt"` // 创建时间
-	UpdatedAt uint32 `field:"updatedAt"` // 修改时间
+	CreatedAt uint64 `field:"createdAt"` // 创建时间
+	UpdatedAt uint64 `field:"updatedAt"` // 修改时间
 	State     uint8  `field:"state"`     // 状态
 }
